requirements.txt
Flask==3.1.0
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
strands-telegram>=0.1.0
reportlab==4.1.0
requests==2.32.3
